moved valuerange code to ffuf main package
