+++ pkg/filter/size.go
@@ -9,13 +9,13 @@ import (
 )
 
 type SizeFilter struct {
-	Value []ValueRange
+	Value []ffuf.ValueRange
 }
 
 func NewSizeFilter(value string) (ffuf.FilterProvider, error) {
-	var intranges []ValueRange
+	var intranges []ffuf.ValueRange
 	for _, sv := range strings.Split(value, ",") {
-		vr, err := ValueRangeFromString(sv)
+		vr, err := ffuf.ValueRangeFromString(sv)
 		if err != nil {
 			return &SizeFilter{}, fmt.Errorf("Size filter or matcher (-fs / -ms): invalid value: %s", sv)
 		}
@@ -27,7 +27,7 @@ func NewSizeFilter(value string) (ffuf.FilterProvider, error) {
 
 func (f *SizeFilter) Filter(response *ffuf.Response) (bool, error) {
 	for _, iv := range f.Value {
-		if iv.min <= response.ContentLength && response.ContentLength <= iv.max {
+		if iv.Min <= response.ContentLength && response.ContentLength <= iv.Max {
 			return true, nil
 		}
 	}
@@ -37,10 +37,10 @@ func (f *SizeFilter) Filter(response *ffuf.Response) (bool, error) {
 func (f *SizeFilter) Repr() string {
 	var strval []string
 	for _, iv := range f.Value {
-		if iv.min == iv.max {
-			strval = append(strval, strconv.Itoa(int(iv.min)))
+		if iv.Min == iv.Max {
+			strval = append(strval, strconv.Itoa(int(iv.Min)))
 		} else {
-			strval = append(strval, strconv.Itoa(int(iv.min))+"-"+strconv.Itoa(int(iv.max)))
+			strval = append(strval, strconv.Itoa(int(iv.Min))+"-"+strconv.Itoa(int(iv.Max)))
 		}
 	}
 	return fmt.Sprintf("Response size: %s", strings.Join(strval, ","))
