+++ pkg/filter/status.go
@@ -11,16 +11,16 @@ import (
 const AllStatuses = 0
 
 type StatusFilter struct {
-	Value []ValueRange
+	Value []ffuf.ValueRange
 }
 
 func NewStatusFilter(value string) (ffuf.FilterProvider, error) {
-	var intranges []ValueRange
+	var intranges []ffuf.ValueRange
 	for _, sv := range strings.Split(value, ",") {
 		if sv == "all" {
-			intranges = append(intranges, ValueRange{AllStatuses, AllStatuses})
+			intranges = append(intranges, ffuf.ValueRange{AllStatuses, AllStatuses})
 		} else {
-			vr, err := ValueRangeFromString(sv)
+			vr, err := ffuf.ValueRangeFromString(sv)
 			if err != nil {
 				return &StatusFilter{}, fmt.Errorf("Status filter or matcher (-fc / -mc): invalid value %s", sv)
 			}
@@ -32,11 +32,11 @@ func NewStatusFilter(value string) (ffuf.FilterProvider, error) {
 
 func (f *StatusFilter) Filter(response *ffuf.Response) (bool, error) {
 	for _, iv := range f.Value {
-		if iv.min == AllStatuses && iv.max == AllStatuses {
+		if iv.Min == AllStatuses && iv.Max == AllStatuses {
 			// Handle the "all" case
 			return true, nil
 		}
-		if iv.min <= response.StatusCode && response.StatusCode <= iv.max {
+		if iv.Min <= response.StatusCode && response.StatusCode <= iv.Max {
 			return true, nil
 		}
 	}
@@ -46,12 +46,12 @@ func (f *StatusFilter) Filter(response *ffuf.Response) (bool, error) {
 func (f *StatusFilter) Repr() string {
 	var strval []string
 	for _, iv := range f.Value {
-		if iv.min == AllStatuses && iv.max == AllStatuses {
+		if iv.Min == AllStatuses && iv.Max == AllStatuses {
 			strval = append(strval, "all")
-		} else if iv.min == iv.max {
-			strval = append(strval, strconv.Itoa(int(iv.min)))
+		} else if iv.Min == iv.Max {
+			strval = append(strval, strconv.Itoa(int(iv.Min)))
 		} else {
-			strval = append(strval, strconv.Itoa(int(iv.min))+"-"+strconv.Itoa(int(iv.max)))
+			strval = append(strval, strconv.Itoa(int(iv.Min))+"-"+strconv.Itoa(int(iv.Max)))
 		}
 	}
 	return fmt.Sprintf("Response status: %s", strings.Join(strval, ","))
